% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{convolve_from_c}
\alias{convolve_from_c}
\title{Call C function from Rcpp}
\usage{
convolve_from_c(a, b)
}
\arguments{
\item{a, b}{A \code{numeric} vector.}
}
\value{
A \code{numeric} vector of length \eqn{N_a + N_b}.
}
\description{
Uses the convolve_c function inside of a C++ routine by Rcpp.
}
\examples{

convolve_from_c(1:5, 5:1)

}
