% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cRcpp-package.R
\docType{package}
\name{cRcpp-package}
\alias{cRcpp}
\alias{cRcpp-package}
\title{cRcpp: Combining a C Routine with Rcpp}
\description{
Provides an implementation of incorporating a 'C' routine into
    an 'Rcpp' routine.
}
\details{
Disclaimer. Work contained within package is part of an R&D effort to
show a minimal example of how something can be achieved. Care should be
used when transferring it into production.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/r-pkg-examples/rcpp-and-fortran}
}

}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu}

}
